# docs/swagger.yaml
basePath: /
definitions:
...
  server_internal_module_system_model.Role:
    properties:
      createdAt:
        type: string
      dataScope:
        type: string
      id:
        type: integer
      isSystem:
        type: integer
      key:
        type: string
      name:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/server_internal_module_system_model.User'
        type: array
    type: object
  server_internal_module_system_model.User:
    properties:
      address:
        type: string
      avatar:
        type: string
      city:
        type: string
      createdAt:
        type: string
      department:
        type: string
      district:
        type: string
      email:
        type: string
      gender:
        type: integer
      id:
        type: integer
      isAdmin:
        type: integer
      jobTitle:
        type: string
      lastLoginAt:
        type: string
      lastLoginIP:
        type: string
      nickname:
        type: string
      phone:
        type: string
      position:
        description: "\U0001F4BC 职业信息"
        type: string
      province:
        description: "\U0001F4CD 地理信息"
        type: string
      roles:
        items:
          $ref: '#/definitions/server_internal_module_system_model.Role'
        type: array
      status:
        type: integer
      tags:
        description: "\U0001F3F7️ 标签信息"
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  server_internal_module_system_model_reply.ApiReply:
    properties:
      createdAt:
        type: string
      description:
        type: string
      group:
        type: string
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      path:
        type: string
      status:
        type: integer
      updatedAt:
        type: string
    type: object
  server_internal_module_system_model_reply.GetUserInfoReply:
    properties:
      address:
        type: string
      avatar:
        type: string
      city:
        type: string
      createdAt:
        description: 注册时间
        type: string
      department:
        type: string
      district:
        type: string
      email:
        type: string
      gender:
        type: integer
      id:
        type: integer
      isAdmin:
        type: integer
      jobTitle:
        type: string
      lastLoginAt:
        type: string
      lastLoginIP:
        type: string
      nickname:
        type: string
      phone:
        type: string
      position:
        type: string
      province:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        type: integer
      tags:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  server_internal_module_system_model_reply.ListApiReply:
    properties:
      list:
        items:
          $ref: '#/definitions/server_internal_module_system_model_reply.ApiReply'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  server_internal_module_system_model_reply.UserListReply:
    properties:
      list:
        items:
          $ref: '#/definitions/server_internal_module_system_model.User'
        type: array
      total:
        type: integer
    type: object
  server_internal_module_system_model_request.CreateApiReq:
    properties:
      description:
        type: string
      group:
        type: string
      method:
        type: string
      name:
        type: string
      path:
        type: string
    required:
    - description
    - group
    - method
    - name
    - path
    type: object
...
  server_internal_module_system_model_request.CreateUserReq:
    properties:
      password:
        type: string
      roleKey:
        items:
          type: string
        type: array
      username:
        type: string
    required:
    - password
    - roleKey
    - username
    type: object
  server_internal_module_system_model_request.DeleteUserReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  server_internal_module_system_model_request.UpdateApiReq:
    properties:
      description:
        type: string
      group:
        type: string
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      path:
        type: string
      status:
        type: integer
    required:
    - id
    - method
    - name
    - path
    type: object
...
host: localhost:9999
info:
  contact:
    email: support@example.com
    name: API Support
  description: 系统管理后台API接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 系统管理API
  version: "1.0"
paths:
...
  /api/system/api:
    post:
      consumes:
      - application/json
      parameters:
      - description: API信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_module_system_model_request.CreateApiReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      security:
      - Bearer: []
      summary: 创建API
      tags:
      - 接口管理
    put:
      consumes:
      - application/json
      parameters:
      - description: API信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_module_system_model_request.UpdateApiReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      security:
      - Bearer: []
      summary: 更新API
      tags:
      - 接口管理
  /api/system/api/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: API ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      security:
      - Bearer: []
      summary: 删除API
      tags:
      - 接口管理
  /api/system/api/list:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: current
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 接口名称
        in: query
        name: name
        type: string
      - description: 接口路径
        in: query
        name: path
        type: string
      - description: 请求方法
        in: query
        name: method
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_module_system_model_reply.ListApiReply'
      security:
      - Bearer: []
      summary: 获取API列表
      tags:
      - 接口管理
...
  /api/system/user:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_module_system_model_request.DeleteUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_module_system_model_request.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      security:
      - Bearer: []
      summary: 创建用户
      tags:
      - 用户管理
  /api/system/user/info:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_module_system_model_reply.GetUserInfoReply'
      security:
      - Bearer: []
      summary: 获取当前用户信息
      tags:
      - 用户管理
  /api/system/user/list:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: current
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_module_system_model_reply.UserListReply'
      security:
      - Bearer: []
      summary: 获取用户列表
      tags:
      - 用户管理
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"